Fix regression causing incorrect effect always-on score lines
Effect always-on score lines have a fractional part so they don't get turned off by CC events. Commit 4c613e1bef7cda0517e79af4e039234ebfe564b1 removed the CONCAT macro used to create the always-on score lines for effects, which caused a space to be inserted between the instrument number and its fractional part. The CONCAT macro was removed in a failed attempt to fix an unrelated error while improving the output folder structure. The error was actually caused by using the wrong include directory, but the CONCAT macro was not added back in again after the actual issue was root-caused and fixed.

This change fixes the regression by adding the CONCAT macro back in again.