Work around channel count limitation in Csound `subinstr` opcode
The Csound `subinstr` opcode can only output a max channel count equal to `nchnls`. This causes an issue in the generated playback .csd because `nchnls` is set to 2 and there are 6 channels used internally and 5 channels being returned by the `subinstr` opcode, which causes a Csound runtime error.

This change works around the issue by updating the instrument signal arrays directly in the sub-instruments when generating the playback .csd.