Fix issue causing missing reverb in playback .csd
When generating the mode 3 trackset score, the CONCAT macro needs to be used when using fractional instrument numbers, otherwise the preprocessor adds a space between the instrument name macro definition and the instrument number's fraction. The lack of a CONCAT macro for the Reverb instrument's always-on score statement caused it to be called with a non-fractional instrument number and an incorrect number of p-fields all offset by 1.

The resulting score statements are:
`i 8 .1 0 -1 1 0` <-- Wrong! There should be no space between the `8` and the `.1`.
`i 8.1 0 -1 1 0`  <-- Correct output fixed in this change by using the CONCAT macro in DawService.csd.