Lower Csound camera matrix update rate to 10 times a second from 30
Csound WASM only gets table updates from BabylonJs about 5 or 6 times a second, so 30 is overkill. 10 is enough to make BabylonJs send a little less than twice as fast as Csound WASM receives. Regardless, 5 or 6 updates a second is not a great update rate and might need to be investigated to see if it can be successfully increased.