Fix issue causing Csound listener position to stop updating after repeat
When repeating, BabylonJs rewinds the score, but since instr 1 has its always on score event in the orc section of the csd, it doesn't get restarted when the score is rewound. This change moves the instr 1 always on score event to the score section of the csd so it gets restarted when the score is rewound and the listener position continues to update after repeats.