Generate new plugin UUIDs when duplicating Reaper tracks
Note that this change now requires the DawService to be restarted before each individual Reaper project is loaded. To make this more automatic, the DawService is now a plugin that should be added to a dummy track in the Reaper project so its UUIDs get reset when the Reaper project is closed and opened. As a result of this workflow change only one Reaper project may be loaded at a time.