Use i-time global variable for k-rate instead of Csound's kr opcode
The kr opcode returns the instrument's local k-rate when set by Csound's setksmps opcode, which makes it unsuitable for use when tracking the time elapsed since the .csd was started.

This change fixes the issue by caching the global k-rate in a global i-rate variable and uses that i-rate variable in the time_i and time_k UDOs instead of Csound's kr opcode.