Force Csound CMake to configure before making
The common Rim mesh .orc file CMake configure step is not run when the Rim mesh .js files change, so the Csound Rim synth mesh data is not always updated by the `node make` command alone.

This change forces the Csound CMake configure step to run before making to fix the issue, which is a temporary solution. A more robust solution would add the .js files as dependencies to a CMake command target for calling CMake configure on the common Rim mesh .orc file.