Make sustain decay use instrument time in Saw1RimSynth, not global time
Saw1RimSynth is using the global time_k UDO for the sustain decay instrument. This causes issues when the score is rewound because the global time is not rewound.

This change fixes the issue by replacing the time_k UDO with the time_NoteTime opcode and updating the envelope calculation accordingly.

Note that this issue was incorrectly blamed on the xadsr opcode in commit cf4c681.