Fix Csound compile error in production build
In the shared modules, `this.constructor.name` is being used when generating the string used in the Csound .csd. This works fine in the development build, but in the production build `this.constructor.name` is empty, resulting in a Csound compile error.

This change fixes the issue by replacing `this.constructor.name` in the shared modules with explicitly set names.